Reorganise modules
Our previous structure did not make sense. We had a state.rs which contained the
storage models at the top level. The actual protocol state was a level lower in
swap::alice and swap::bob. We had protocol code (negotiate.rs/steps.rs) in the
same module as our stateless network code. Created 3 modules for protocol,
network and storage and put the files where they should belong.
