+++ swap/src/lib.rs
@@ -3,14 +3,12 @@
 use ::serde::{Deserialize, Serialize};
 use std::fmt::{self, Display};
 
-pub mod alice;
 pub mod bitcoin;
-pub mod bob;
 pub mod cli;
 pub mod monero;
 pub mod network;
+pub mod protocol;
 pub mod serde;
-pub mod state;
 pub mod storage;
 pub mod tor;
 pub mod trace;
