+++ swap/src/protocol.rs
@@ -1,0 +1,2 @@
+pub mod alice;
+pub mod bob;
