+++ swap/src/storage.rs
@@ -1,8 +1,10 @@
-use crate::state::Swap;
 use anyhow::{anyhow, bail, Context, Result};
+use model::Swap;
 use serde::{de::DeserializeOwned, Serialize};
 use std::path::Path;
 use uuid::Uuid;
+
+pub mod model;
 
 pub struct Database(sled::Db);
 
@@ -81,7 +83,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::state::{Alice, Bob};
+    use crate::storage::model::{Alice, Bob};
 
     use super::*;
